Сделал "job" для добавления лайков на посты

+++ backend-laravel/database/migrations/2022_12_29_161440_create_accounts_table.php
@@ -15,7 +15,7 @@ class CreateAccountsTable extends Migration
     {
         Schema::create('accounts', function (Blueprint $table) {
             $table->id();
-            $table->integer('account_id')->unsigned()->unique();
+            $table->bigInteger('account_id')->unsigned()->unique();
             $table->string('access_token');
             $table->string('screen_name');
             $table->string('first_name');
